requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
mysql-connector-python==8.4.0
openai==1.35.14
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
uvicorn==0.30.1
openai-whisper==20231117
webvtt-py==0.5.1
tqdm==4.66.4
SpeechRecognition==3.10.4
moviepy==1.0.3
googletrans-py==4.0.0